refactor: 검색어 도메인 위치 변경
