+++ src/main/java/com/recipe/app/src/recipe/infra/keyword/SearchKeywordRepository.java
@@ -1,6 +1,6 @@
 package com.recipe.app.src.recipe.infra.keyword;
 
-import com.recipe.app.src.recipe.domain.SearchKeyword;
+import com.recipe.app.src.recipe.domain.keyword.SearchKeyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
